re-structure exports in chain.rs (make modules public);
extract tx signing into signing.rs;
extract ErgoStateContext;
