+++ sigma-tree/src/chain.rs
@@ -1,30 +1,25 @@
 //! Ergo chain types
 
-mod address;
-mod base16_bytes;
-mod box_id;
-mod context_extension;
-mod contract;
-mod data_input;
-mod digest32;
-mod ergo_box;
-mod input;
 #[cfg(feature = "with-serde")]
 mod json;
-mod prover_result;
-mod token;
-mod transaction;
 
-pub use address::*;
+mod base16_bytes;
+// TODO: move to ergo_box
+mod box_id;
+mod digest32;
+
 pub use base16_bytes::Base16DecodedBytes;
 pub use base16_bytes::Base16EncodedBytes;
 pub use box_id::*;
-pub use context_extension::*;
-pub use contract::*;
 pub use digest32::*;
-pub use ergo_box::*;
-pub use input::*;
-pub use prover_result::*;
-pub use prover_result::*;
-pub use token::*;
-pub use transaction::*;
+
+pub mod address;
+pub mod context_extension;
+pub mod contract;
+pub mod data_input;
+pub mod ergo_box;
+pub mod ergo_state_context;
+pub mod input;
+pub mod prover_result;
+pub mod token;
+pub mod transaction;
