+++ sigma-tree/src/chain/ergo_state_context.rs
@@ -1,0 +1,4 @@
+//! Blockchain state
+
+/// Blockchain state (last headers, etc.)
+pub struct ErgoStateContext();
