+++ sigma-tree/src/ergo_tree.rs
@@ -299,12 +299,12 @@ mod tests {
 
     #[test]
     fn test_constant_segregation_header_flag_support() {
-        let encoder = chain::AddressEncoder::new(chain::NetworkPrefix::Mainnet);
+        let encoder = chain::address::AddressEncoder::new(chain::address::NetworkPrefix::Mainnet);
         let address = encoder
             .parse_address_from_str("9hzP24a2q8KLPVCUk7gdMDXYc7vinmGuxmLp5KU7k9UwptgYBYV")
             .unwrap();
 
-        let contract = chain::Contract::pay_to_address(address).unwrap();
+        let contract = chain::contract::Contract::pay_to_address(address).unwrap();
         let bytes = &contract.get_ergo_tree().sigma_serialise_bytes();
         assert_eq!(&bytes[..2], vec![0u8, 8u8].as_slice());
     }
